requirements.txt
blas=1.0=mkl
ca-certificates=2021.10.8=h033912b_0
certifi=2021.5.30=py36h79c6626_0
greenlet=1.1.1=py36h23ab428_0
importlib-metadata=4.8.1=py36hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
krb5=1.19.2=hcd88c3b_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libpq=12.2=h1b4eb34_1
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.3.0=py36ha059aab_0
mkl_random=1.1.1=py36h959d312_0
ncurses=6.3=hca72f7f_2
numpy=1.19.2=py36h456fd55_0
numpy-base=1.19.2=py36hcfb5961_0
openssl=1.1.1l=h0d85af4_0
pandas=1.1.5=py36hb2f4e1b_0
pip=21.2.2=py36hecd8cb5_0
psycopg2=2.8.6=py36hbcfaee0_1
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-decouple=3.4=pyhd3deb0d_0
python_abi=3.6=2_cp36m
pytz=2021.3=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
setuptools=58.0.4=py36hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlalchemy=1.4.22=py36h9ed2024_0
sqlite=3.36.0=hce871da_0
tk=8.6.11=h7bc2e8c_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
pandas~=1.1.5
